pyproject.toml
[tool.poetry]
name = "cryptoversion"
version = "0.1.0"
description = "CLI application to convert cryptocurrencies into fiat"
authors = ["datiswaken <jannis.lambinus@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dependency-injector = "^4.41.0"
requests = "^2.31.0"
pyyaml = "^6.0.1"
polars = "^0.20.15"
numpy = "^1.26.4"


[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
types-requests = "^2.31.0.20240406"
types-pyyaml = "^6.0.12.20240311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
